Reduce amount of code in headers
