+++ src/CMakeLists.txt
@@ -1,3 +1,8 @@
+# Add sources
+target_sources(ChessPlusPlus PUBLIC
+	"${CMAKE_CURRENT_LIST_DIR}/Debug.cpp"
+)
+
 # Divvy out work to subdirectories
 add_subdirectory("app/")
 add_subdirectory("board/")
