+++ src/config/CMakeLists.txt
@@ -1,4 +1,5 @@
 # Add sources
 target_sources(ChessPlusPlus PUBLIC
+	"${CMAKE_CURRENT_LIST_DIR}/BoardConfig.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/Configuration.cpp"
 )
