+++ src/config/Configuration.cpp
@@ -1,6 +1,13 @@
 #include "Configuration.hpp"
 
+#include "Exception.hpp"
+
 #include <iostream>
+#include <fstream>
+#include <cstring>
+#include <cstdint>
+#include <boost/algorithm/string/replace.hpp>
+#include <boost/filesystem.hpp>
 
 #if defined(__linux__)
 #include <unistd.h>
@@ -53,6 +60,29 @@ namespace chesspp
             std::clog << "Working directory = \"" << boost::filesystem::current_path().string() << '"' << std::endl;
 
             return ret;
+        }
+        std::string Configuration::validateConfigFile(std::string const &configFile)
+        {
+            static std::string exe_path = executablePath();
+
+            if(boost::filesystem::extension(configFile) != ".json")
+            {
+                throw Exception("Configuration cannot read non-json config files.");
+            }
+
+            if(boost::filesystem::exists(configFile))
+            {
+                res_path = "";
+            }
+            else
+            {
+                res_path = exe_path;
+            }
+            return res_path + configFile;
+        }
+        Configuration::Configuration(std::string const &configFile) noexcept(false)
+        : reader{std::ifstream(validateConfigFile(configFile))}
+        {
         }
     }
 }
