+++ src/gfx/CMakeLists.txt
@@ -1,4 +1,4 @@
 # Add sources
 target_sources(ChessPlusPlus PUBLIC
-	"${CMAKE_CURRENT_LIST_DIR}/Graphics.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/GraphicsHandler.cpp"
 )
