+++ src/piece/Archer.hpp
@@ -1,5 +1,5 @@
-#ifndef ChessPlusPlus_Piece_ArcherChessPiece_HeaderPlusPlus
-#define ChessPlusPlus_Piece_ArcherChessPiece_HeaderPlusPlus
+#ifndef chesspp_piece_Archer_HeaderPlusPlus
+#define chesspp_piece_Archer_HeaderPlusPlus
 
 #include "board/Board.hpp"
 #include "piece/Piece.hpp"
@@ -8,7 +8,8 @@ namespace chesspp
 {
     namespace piece
     {
-        class Archer : public virtual Piece
+        class Archer
+        : public virtual Piece
         {
         public:
             Archer(board::Board &b, Position_t const &pos, Suit_t const &s, Class_t const &pc);
