+++ src/piece/Bishop.hpp
@@ -1,5 +1,5 @@
-#ifndef ChessPlusPlus_Piece_BishopChessPiece_HeaderPlusPlus
-#define ChessPlusPlus_Piece_BishopChessPiece_HeaderPlusPlus
+#ifndef chesspp_piece_Bishop_HeaderPlusPlus
+#define chesspp_piece_Bishop_HeaderPlusPlus
 
 #include "board/Board.hpp"
 #include "piece/Piece.hpp"
@@ -8,7 +8,8 @@ namespace chesspp
 {
     namespace piece
     {
-        class Bishop : public virtual Piece
+        class Bishop
+        : public virtual Piece
         {
         public:
             Bishop(board::Board &b, Position_t const &pos, Suit_t const &s, Class_t const &pc);
