+++ src/piece/King.hpp
@@ -1,5 +1,5 @@
-#ifndef ChessPlusPlus_Piece_KingChessPiece_HeaderPlusPlus
-#define ChessPlusPlus_Piece_KingChessPiece_HeaderPlusPlus
+#ifndef chesspp_piece_King_HeaderPlusPlus
+#define chesspp_piece_King_HeaderPlusPlus
 
 #include "board/Board.hpp"
 #include "piece/Piece.hpp"
@@ -8,7 +8,8 @@ namespace chesspp
 {
     namespace piece
     {
-        class King : public virtual Piece
+        class King
+        : public virtual Piece
         {
         public:
             King(board::Board &b, Position_t const &pos, Suit_t const &s, Class_t const &pc);
