+++ src/piece/Knight.hpp
@@ -1,5 +1,5 @@
-#ifndef ChessPlusPlus_Piece_KnightChessPiece_HeaderPlusPlus
-#define ChessPlusPlus_Piece_KnightChessPiece_HeaderPlusPlus
+#ifndef chesspp_piece_Knight_HeaderPlusPlus
+#define chesspp_piece_Knight_HeaderPlusPlus
 
 #include "board/Board.hpp"
 #include "piece/Piece.hpp"
@@ -8,7 +8,8 @@ namespace chesspp
 {
     namespace piece
     {
-        class Knight : public virtual Piece
+        class Knight
+        : public virtual Piece
         {
         public:
             Knight(board::Board &b, Position_t const &pos, Suit_t const &s, Class_t const &pc);
