+++ src/piece/Queen.hpp
@@ -1,5 +1,5 @@
-#ifndef ChessPlusPlus_Piece_QueenChessPiece_HeaderPlusPlus
-#define ChessPlusPlus_Piece_QueenChessPiece_HeaderPlusPlus
+#ifndef chesspp_piece_Queen_HeaderPlusPlus
+#define chesspp_piece_Queen_HeaderPlusPlus
 
 #include "board/Board.hpp"
 #include "piece/Piece.hpp"
@@ -8,7 +8,8 @@ namespace chesspp
 {
     namespace piece
     {
-        class Queen : public virtual Piece
+        class Queen
+        : public virtual Piece
         {
         public:
             Queen(board::Board &b, Position_t const &pos, Suit_t const &s, Class_t const &pc);
