+++ src/piece/Rook.hpp
@@ -1,5 +1,5 @@
-#ifndef ChessPlusPlus_Piece_RookChessPiece_HeaderPlusPlus
-#define ChessPlusPlus_Piece_RookChessPiece_HeaderPlusPlus
+#ifndef chesspp_piece_Rook_HeaderPlusPlus
+#define chesspp_piece_Rook_HeaderPlusPlus
 
 #include "board/Board.hpp"
 #include "piece/Piece.hpp"
@@ -8,7 +8,8 @@ namespace chesspp
 {
     namespace piece
     {
-        class Rook : public virtual Piece
+        class Rook
+        : public virtual Piece
         {
         public:
             Rook(board::Board &b, Position_t const &pos, Suit_t const &s, Class_t const &pc);
