+++ src/util/CMakeLists.txt
@@ -1,0 +1,5 @@
+# Add sources
+target_sources(ChessPlusPlus PUBLIC
+	"${CMAKE_CURRENT_LIST_DIR}/Direction.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/JsonReader.cpp"
+)
