+++ src/util/JsonReader.hpp
@@ -1,16 +1,15 @@
-#ifndef ChessPlusPlus_Util_JsonReaderClass_HeaderPlusPlus
-#define ChessPlusPlus_Util_JsonReaderClass_HeaderPlusPlus
+#ifndef chesspp_util_JsonReader_HeaderPlusPlus
+#define chesspp_util_JsonReader_HeaderPlusPlus
 
-#include "Exception.hpp"
-
-#include <json.h>
-
+#include <memory>
 #include <string>
-#include <istream>
+#include <iosfwd>
 #include <streambuf>
 #include <cstdint>
 #include <map>
 #include <utility>
+
+struct _json_value;
 
 namespace chesspp
 {
@@ -21,12 +20,12 @@ namespace chesspp
          * Current implementation uses json-parser by
          * James McLaughlin. Thus, it wraps a json_value.
          */
-        class JsonReader
+        class JsonReader final
         {
             /**
              * Underlying json_value pointer.
              */
-            json_value *json {nullptr};
+            _json_value *json {nullptr};
         public:
             JsonReader() = delete;
             JsonReader(JsonReader const &) = delete;
@@ -36,66 +35,45 @@ namespace chesspp
              * The stream is read to EOF.
              * \param s The stream containing the JSON.
              */
-            JsonReader(std::istream &s)
-            {
-                if(!s)
-                {
-                    throw Exception("stream given to JsonReader in bad state");
-                }
-                std::string str ((std::istreambuf_iterator<char>(s)), std::istreambuf_iterator<char>());
-                json_settings options {0, 0, nullptr, nullptr, nullptr, 0U};
-                char error[json_error_max];
-                json = json_parse_ex(&options, str.c_str(), str.length(), error);
-                if(json == nullptr)
-                {
-                    //no manual cleanup needed
-                    throw Exception(std::string("Error loading JSON: ") + error);
-                }
-            }
+            JsonReader(std::istream &s);
             /**
              * Constructs this JsonReader from the given temporary stream.
              * The stream is read to EOF.
              * \param s The temporary stream containing the JSON.
              */
-            JsonReader(std::istream &&s)
-            : JsonReader{s}
-            {
-            }
+            JsonReader(std::istream &&s);
             /**
              * Move-constructs this JsonReader from another.
              * \param from the JsonReader to move.
              */
-            JsonReader(JsonReader &&from)
-            : json{from.json}
-            {
-                from.json = nullptr;
-            }
+            JsonReader(JsonReader &&from);
             /**
              * Move-assigns this JsonReader from a given stream rvalue.
              * The stream is read to EOF. Returns *this
              * \param s The stream containing the JSON.
              */
-            JsonReader &operator=(std::istream &&s)
-            {
-                *this = JsonReader(s); //move assign
-                return *this;
-            }
+            JsonReader &operator=(std::istream &&s);
             /**
              * Move-assigns this JsonReader from another.
              * \param from the JsonReader to move.
              */
-            JsonReader &operator=(JsonReader &&from)
-            {
-                std::swap(json, from.json);
-                return *this;
-            }
+            JsonReader &operator=(JsonReader &&from);
             /**
              * Destructs this JsonReader freeing any allocated memory.
              */
-            ~JsonReader()
+            ~JsonReader();
+
+            enum class Type
             {
-                json_value_free(json), json = nullptr;
-            }
+                None,
+                Object,
+                Array,
+                Integer,
+                Double,
+                String,
+                Boolean,
+                Null
+            };
 
             /**
              * Represents a value in the JSON.
@@ -113,17 +91,14 @@ namespace chesspp
                 /**
                  * The particular value this instance is associated with.
                  */
-                json_value const &value;
+                _json_value const &value;
                 /**
                  * Constructs this NestedValue from a json_value.
                  * Note that this constructor is private and for
                  * implementation use only.
                  * \param value_ The json_value this instance shall represent.
                  */
-                NestedValue(json_value const &value_) noexcept
-                : value(value_) //can't use {}
-                {
-                }
+                NestedValue(_json_value const &) noexcept;
                 NestedValue &operator=(NestedValue const &) noexcept = delete;
             public:
                 NestedValue(NestedValue const &) = default;
@@ -145,24 +120,14 @@ namespace chesspp
                  * json_object
                  * \return The type of value represented.
                  */
-                json_type type() const noexcept //may wish to add an abstraction layer between json-parser
-                {
-                    return value.type;
-                }
+                Type type() const noexcept;
                 /**
                  * Returns the parent NestedValue or throws
-                 * ::chesspp::Exception.
+                 * ::std::domain_error.
                  * \return the parent NestedValue.
-                 * \throws ::chesspp::Exception
+                 * \throws ::std::domain_error
                  */
-                NestedValue parent() const
-                {
-                    if(value.parent)
-                    {
-                        return *value.parent;
-                    }
-                    throw Exception("No parent json value");
-                }
+                NestedValue parent() const;
 
                 /**
                  * Returns a NestedValue within this one by name.
@@ -170,42 +135,26 @@ namespace chesspp
                  * \param name the name of the nested JSON value.
                  * \return the NestedValue by name.
                  */
-                NestedValue operator[](std::string const &name) const noexcept(noexcept(name.c_str()))
-                {
-                    return value[name.c_str()];
-                }
+                NestedValue operator[](std::string const &name) const noexcept(noexcept(name.c_str()));
                 /**
                  * Returns a NestedValue within this one by name.
                  * Only works if type() == json_object
                  * \param name the name of the nested JSON value.
                  * \return the NestedValue by name.
                  */
-                NestedValue operator[](char const *name) const //without this, ambiguity occurs
-                {
-                    return value[name];
-                }
+                NestedValue operator[](char const *name) const noexcept;
                 /**
                  * Returns the length of array values.
                  * \return The length of the array, or 0 if not an array.
                  */
-                std::size_t length() const noexcept
-                {
-                    if(value.type == json_array)
-                    {
-                        return value.u.array.length;
-                    }
-                    return 0;
-                }
+                std::size_t length() const noexcept;
                 /**
                  * Returns the nested value at the given array index.
                  * Only works if type() == json_array
                  * \param index The index within the array.
                  * \return the nested value at the given index.
                  */
-                NestedValue operator[](std::size_t index) const noexcept
-                {
-                    return value[static_cast<int>(index)];
-                }
+                NestedValue operator[](std::size_t index) const noexcept;
                 /**
                  * Provides a std::map-based view of
                  * an object value, mapping object keys
@@ -213,61 +162,40 @@ namespace chesspp
                  * type() == json_object
                  * \return a std::map of object keys ot obejct values
                  */
-                std::map<std::string, NestedValue> object() const
-                {
-                    std::map<std::string, NestedValue> obj;
-                    if(value.type == json_object)
-                    {
-                        for(std::size_t i = 0; i < value.u.object.length; ++i)
-                        {
-                            obj.insert(std::make_pair<std::string const, NestedValue>(value.u.object.values[i].name, *value.u.object.values[i].value));
-                        }
-                    }
-                    return obj;
-                }
+                std::map<std::string, NestedValue> object() const;
                 /**
                  * Returns the string representation of this string value.
                  * Only works if type() == json_string
                  * \return The string
                  */
-                operator std::string() const noexcept(noexcept(std::string("")))
-                {
-                    return static_cast<char const *>(value);
-                }
-                //I tried an approach with templates, but the compiler was never able to deduce the corect template argument
-                operator std::  int8_t() const noexcept { return static_cast<std::  int8_t>(static_cast<json_int_t>(value)); }
-                operator std:: uint8_t() const noexcept { return static_cast<std:: uint8_t>(static_cast<json_int_t>(value)); }
-                operator std:: int16_t() const noexcept { return static_cast<std:: int16_t>(static_cast<json_int_t>(value)); }
-                operator std::uint16_t() const noexcept { return static_cast<std::uint16_t>(static_cast<json_int_t>(value)); }
-                operator std:: int32_t() const noexcept { return static_cast<std:: int32_t>(static_cast<json_int_t>(value)); }
-                operator std::uint32_t() const noexcept { return static_cast<std::uint32_t>(static_cast<json_int_t>(value)); }
-                operator std:: int64_t() const noexcept { return static_cast<std:: int64_t>(static_cast<json_int_t>(value)); }
-                operator std::uint64_t() const noexcept { return static_cast<std::uint64_t>(static_cast<json_int_t>(value)); }
+                operator std::string() const noexcept(noexcept(std::string("")));
+                operator std::  int8_t() const noexcept;
+                operator std:: uint8_t() const noexcept;
+                operator std:: int16_t() const noexcept;
+                operator std::uint16_t() const noexcept;
+                operator std:: int32_t() const noexcept;
+                operator std::uint32_t() const noexcept;
+                operator std:: int64_t() const noexcept;
+                operator std::uint64_t() const noexcept;
                 /**
                  * Returns the boolean state of this bool value.
                  * Only works if type() == json_bool.
                  * \return The boolean.
                  */
-                operator bool() const noexcept
-                {
-                    return value;
-                }
+                operator bool() const noexcept;
                 /**
                  * Returns the floating-point value of this double.
                  * Only works if type() == json_double
                  * \return The double.
                  */
-                operator double() const noexcept
-                {
-                    return value;
-                }
+                operator double() const noexcept;
 
                 /**
                  * Only for extreme use cases: returns the
                  * underlying json_value being wrapped.
                  * \return the underlying json_value
                  */
-                json_value const &implementation() noexcept
+                _json_value const &implementation() noexcept
                 {
                     return value;
                 }
@@ -278,10 +206,7 @@ namespace chesspp
              * this JSON.
              * \return a NestedValue view into this JSON.
              */
-            NestedValue access() const noexcept
-            {
-                return *json;
-            }
+            NestedValue access() const noexcept;
             /**
              * Returns a NestedValue view into
              * this JSON.
